Added modes to build and run seperately
`./run.sh build` builds the docker image.
`./run.sh run` run the built image.

The logic behind this is that you'd hardly need to
rebuild your image, as it is not so common to add new
Gems. While it is common to run Jekyll quite often.
Changes to the `md` files will be reflected without
re-building or even re-running.